Downgrade ASP.NET MVC to lowest version
This way it should be compatible with any version of ASP.NET MVC the user
may have installed, since NuGet sets up a binding redirect for all earlier
versions up to the current installed version.

I also recombined the two .dlls into one.  As far as I understand it now,
assembly binding is delayed till a method is called that references
something in the other assembly.