Add dynamic MVC handler locator implementation
While the current version of VS/NuGet appears to set up a binding redirect
for MVC that should allow for the static locator to work, I have seen
existing projects where this is not the case.  For maximum compatibility,
I've added a second implementation that uses reflection to take advantage
of any version of MVC that may be available.

I've left both implementations active. The static implementation is used
where available for performance reasons, while the dynamic implementation
is used as a fallback.

In theory, an approach that's an alternateive to using reflection would be
to handle the AppDomain.AssemblyResolve event and set up the binding
redirect in code.  For whatever reason, I could not get the event handler
to fire in an HttpModule despite trying many approaches, so that's why I'm
not using it.